Restructure tcan_can
