+++ tcan_can/src/DeviceJ1939.cpp
@@ -1,0 +1,7 @@
+#include "tcan_can/DeviceJ1939.hpp"
+
+#include "tcan_can/CanBus.hpp"
+
+namespace tcan_can {
+
+}  // namespace tcan_can
